+++ tests/cpu/seq_single_module.cpp
@@ -59,7 +59,7 @@ TEST(algorithms, seq_single_module) {
         det_desc.bin_edges_y()[i] = std::move(bin_edges_y);
         det_desc.dimensions()[i] = 2;
         det_desc.subspace()[i]  = {0, 1};
-        det_desc.design_id()[i] = static_cast<float>(i);
+        det_desc.design_id()[i] = static_cast<int>(i);
     }
 
     auto cells_data = vecmem::get_data(cells);
